moved back the trait implementation in liquidity pallet
