+++ pallets/liquidity/Cargo.toml
@@ -21,8 +21,6 @@ polkadex-primitives = { workspace = true, default-features = false }
 frame-benchmarking = { workspace = true, default-features = false, optional = true }
 pallet-timestamp = { workspace = true, default-features = false }
 test-utils = { workspace = true, default-features = false, optional = true }
-thea-primitives = { path = "../../thea-primitives", default-features = false }
-pallet-ocex-lmp = {path = "../ocex", default-features = false}
 
 [dev-dependencies]
 test-utils = { workspace = true }
@@ -33,7 +31,7 @@ sp-keystore = { workspace = true }
 frame-benchmarking = { workspace = true }
 anyhow = { workspace = true }
 sp-io = { workspace = true }
-
+pallet-ocex-lmp = { path = "../ocex", default-features = false }
 
 [features]
 default = ["std"]
@@ -48,15 +46,13 @@ std = [
     "polkadex-primitives/std",
     "pallet-timestamp/std",
     "pallet-ocex-lmp/std",
-    "thea-primitives/std",
 ]
 runtime-benchmarks = [
     "sp-runtime/runtime-benchmarks",
     "frame-benchmarking/runtime-benchmarks",
     "frame-support/runtime-benchmarks",
     "frame-system/runtime-benchmarks",
-    "pallet-ocex-lmp/runtime-benchmarks",
     "test-utils",
-    "thea-primitives/runtime-benchmarks",
+    "pallet-ocex-lmp/runtime-benchmarks",
 ]
 try-runtime = ["frame-support/try-runtime"]
