+++ pallets/liquidity/src/benchmarking.rs
@@ -19,14 +19,13 @@
 //! Benchmarking setup for liquidity pallet
 #![cfg(feature = "runtime-benchmarks")]
 use super::*;
-use crate::{pallet::Call, Pallet as liquidity};
+use crate::{pallet::Call, LiquidityModifier, Pallet as liquidity};
 use frame_benchmarking::benchmarks;
 use frame_support::{dispatch::UnfilteredDispatchable, traits::EnsureOrigin};
 use frame_system::RawOrigin;
 use parity_scale_codec::Decode;
 use polkadex_primitives::{AssetId, UNIT_BALANCE};
 use sp_runtime::SaturatedConversion;
-use thea_primitives::liquidity::LiquidityModifier;
 
 // Check if last event generated by pallet is the one we're expecting
 fn assert_last_event<T: Config>(generic_event: <T as Config>::Event) {
