+++ pallets/ocex/Cargo.toml
@@ -27,7 +27,7 @@ rust_decimal = { workspace = true, features = ["scale-codec"], default-features 
 frame-benchmarking = { workspace = true, default-features = false, optional = true }
 test-utils = { workspace = true, default-features = false, optional = true }
 pallet-timestamp = { workspace = true, default-features = false }
-thea-primitives = { path = "../../thea-primitives", default-features = false }
+liquidity = { path = "../liquidity", default-features = false }
 
 [dev-dependencies]
 test-utils = { workspace = true }
@@ -54,8 +54,7 @@ std = [
     "ias-verify/std",
     "rust_decimal/std",
     "pallet-timestamp/std",
-#    "liquidity/std",
-    "thea-primitives/std"
+    "liquidity/std",
 ]
 runtime-benchmarks = [
     "sp-runtime/runtime-benchmarks",
@@ -63,6 +62,6 @@ runtime-benchmarks = [
     "frame-support/runtime-benchmarks",
     "frame-system/runtime-benchmarks",
     "test-utils",
-    "thea-primitives/runtime-benchmarks"
+    "liquidity/runtime-benchmarks",
 ]
 try-runtime = ["frame-support/try-runtime"]
